+++ model/resources.go
@@ -7,8 +7,12 @@ type Rand struct {
 	rand.Source // Source to use for PRNGs in [System] implementations.
 }
 
-// Time is a resource holding the model's time step.
-type Time struct {
-	Tick     int64 // The current model tick.
-	Finished bool  // Whether the model run is finished. Can be set by systems.
+// Tick is a resource holding the model's time step.
+type Tick struct {
+	Tick int64 // The current model tick.
+}
+
+// Termination is a resource holding a whether ths system should terminate after the current step.
+type Termination struct {
+	Terminate bool // Whether the model run is finished. Can be set by systems.
 }