+++ system/callback_termination.go
@@ -9,23 +9,28 @@ import (
 // CallbackTermination system.
 //
 // Terminates a model run according to the return value of a callback function.
+//
+// Expects a resource of type [model.Termination].
 type CallbackTermination struct {
 	Callback func(t int64) bool // The callback. ends the simulation when it returns true.
-	timeRes  generic.Resource[model.Time]
+	termRes  generic.Resource[model.Termination]
+	step     int64
 }
 
 // Initialize the system
 func (s *CallbackTermination) Initialize(w *ecs.World) {
-	s.timeRes = generic.NewResource[model.Time](w)
+	s.termRes = generic.NewResource[model.Termination](w)
+	s.step = 0
 }
 
 // Update the system
 func (s *CallbackTermination) Update(w *ecs.World) {
-	time := s.timeRes.Get()
+	term := s.termRes.Get()
 
-	if s.Callback(time.Tick) {
-		time.Finished = true
+	if s.Callback(s.step) {
+		term.Terminate = true
 	}
+	s.step++
 }
 
 // Finalize the system