+++ system/fixed_termination.go
@@ -9,23 +9,28 @@ import (
 // FixedTermination system.
 //
 // Terminates a model run after a fixed number of ticks.
+//
+// Expects a resource of type [model.Termination].
 type FixedTermination struct {
 	Steps   int64 // Number of simulation ticks to run.
-	timeRes generic.Resource[model.Time]
+	termRes generic.Resource[model.Termination]
+	step    int64
 }
 
 // Initialize the system
 func (s *FixedTermination) Initialize(w *ecs.World) {
-	s.timeRes = generic.NewResource[model.Time](w)
+	s.termRes = generic.NewResource[model.Termination](w)
+	s.step = 0
 }
 
 // Update the system
 func (s *FixedTermination) Update(w *ecs.World) {
-	time := s.timeRes.Get()
+	term := s.termRes.Get()
 
-	if time.Tick+1 >= s.Steps {
-		time.Finished = true
+	if s.step+1 >= s.Steps {
+		term.Terminate = true
 	}
+	s.step++
 }
 
 // Finalize the system