+++ system/fixed_termination_test.go
@@ -16,7 +16,7 @@ func TestFixedTermination(t *testing.T) {
 
 	m.Run()
 
-	time := ecs.GetResource[model.Time](&m.World)
+	time := ecs.GetResource[model.Tick](&m.World)
 	assert.Equal(t, 100, int(time.Tick))
 }