+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [[v0.0.2]](https://github.com/mlange-42/arche/compare/v0.0.1...v0.0.2)
 
+### Breaking changes
+
+* Resource `Time` is now split into `Tick` and `Termination` (#16)
+
 ### Features
 
 * Adds a system `CallbackTermination` to end the simulation based on a callback (#13)
@@ -13,6 +17,7 @@
 ### Other
 
 * Systems are removed immediately when `Systems.RemoveSystem` is called outside of a loop over systems (#15)
+* Included systems do no longer depend on resource `Tick` (formerly `Time`) (#16)
 
 ### Documentation